fix(w2ner): 修复w2ner更新产生的bug
